test: consolidate pipeline test scaffold and reorganize files by concern
Moving tests from badly-named files (coverage, final, additional, advanced)
into concern-based files (pipeline, stages_engine, stages_game, etc).

Deduplication:
- Created setupTestContext() helper to replace repeated FakeEngineTree + config + project setup (9 call sites)
- Created newTestPipelineCtx() functional-option-style helper to reduce pipelineCtx literal boilerplate (10 call sites)
- Moved newTestCache() duplicate from stages_logic_test.go to helpers_test.go (now single definition)

File reorganization:
- Deleted stages_coverage_test.go, stages_final_test.go, pipeline_additional_test.go, stages_game_advanced_test.go
- Consolidated tests into concern-based files: pipeline_test, stages_engine, stages_game, stages_deploy, stages_cache, stages_dispatch, stages_deploy_container, stages_test
- Reduced FakeEngineTree call sites from 20 to 13 through use of setupTestContext helper

Constraints maintained:
- Coverage: 87.9% (unchanged)
- All tests pass
- No function exceeds cyclomatic complexity 8 (verified with gocyclo)
- All files under 500 lines
- No production code modified

Removed duplicate helper:
- Deleted newTestCache() from stages_logic_test.go (consolidated into helpers_test.go)
