+++ cmd/pipeline/pipeline_test.go
@@ -2,6 +2,7 @@ package pipeline
 
 import (
 	"context"
+	"fmt"
 	"strings"
 	"testing"
 
@@ -137,5 +138,136 @@ func TestRunPipelineValidationError(t *testing.T) {
 	}
 	if !strings.Contains(err.Error(), "failed") {
 		t.Errorf("expected error message to contain 'failed', got: %v", err)
+	}
+}
+
+func TestNewPipelineCtxSuccess(t *testing.T) {
+	engineRoot, projectPath, cfg := setupTestContext(t, "TestGame")
+
+	cfg.AWS = config.AWSConfig{
+		AccountID:     "123456789012",
+		Region:        "us-east-1",
+		ECRRepository: "test-game",
+	}
+
+	globals.SetGlobals(t, cfg)
+
+	globals.SwapResolveTarget(t, func(ctx context.Context, c *config.Config, s string) (deploy.Target, error) {
+		return &stubTarget{name: "binary"}, nil
+	})
+
+	cmd := Cmd
+
+	pctx, err := newPipelineCtx(cmd)
+	if err != nil {
+		t.Fatalf("newPipelineCtx() error = %v, want nil", err)
+	}
+
+	if pctx.cfg == nil {
+		t.Error("newPipelineCtx() cfg is nil")
+	}
+	if pctx.r == nil {
+		t.Error("newPipelineCtx() r is nil")
+	}
+	if pctx.target == nil {
+		t.Error("newPipelineCtx() target is nil")
+	}
+
+	_ = engineRoot
+	_ = projectPath
+}
+
+func TestNewPipelineCtxResolveTargetError(t *testing.T) {
+	cfg := &config.Config{
+		Engine: config.EngineConfig{
+			SourcePath: "C:/ue5",
+			Version:    "5.7.3",
+		},
+		Game: config.GameConfig{
+			ProjectName: "TestGame",
+		},
+	}
+
+	globals.SetGlobals(t, cfg)
+
+	globals.SwapResolveTarget(t, func(ctx context.Context, c *config.Config, s string) (deploy.Target, error) {
+		return nil, fmt.Errorf("target resolution failed")
+	})
+
+	cmd := Cmd
+
+	_, err := newPipelineCtx(cmd)
+	if err == nil {
+		t.Fatal("newPipelineCtx() expected error, got nil")
+	}
+}
+
+func TestPrintNextStep(t *testing.T) {
+	// Verify printNextStep with default flags (no session, no deploy skip)
+	orig := skipDeploy
+	origSession := withSession
+	t.Cleanup(func() {
+		skipDeploy = orig
+		withSession = origSession
+	})
+
+	skipDeploy = false
+	withSession = false
+
+	output := captureStdout(printNextStep)
+	if !strings.Contains(output, "Pipeline complete") {
+		t.Errorf("expected 'Pipeline complete' in output, got: %q", output)
+	}
+	if !strings.Contains(output, "ludus deploy session") {
+		t.Errorf("expected 'ludus deploy session' guidance, got: %q", output)
+	}
+	if strings.Contains(output, "ludus connect") {
+		t.Errorf("should not show 'ludus connect' when withSession is false, got: %q", output)
+	}
+}
+
+func TestPrintNextStepWithConnection(t *testing.T) {
+	// Verify printNextStep with session enabled shows connect command
+	orig := skipDeploy
+	origSession := withSession
+	t.Cleanup(func() {
+		skipDeploy = orig
+		withSession = origSession
+	})
+
+	skipDeploy = false
+	withSession = true
+
+	output := captureStdout(printNextStep)
+	if !strings.Contains(output, "Pipeline complete") {
+		t.Errorf("expected 'Pipeline complete' in output, got: %q", output)
+	}
+	if !strings.Contains(output, "ludus connect") {
+		t.Errorf("expected 'ludus connect' guidance, got: %q", output)
+	}
+	if strings.Contains(output, "ludus deploy session") {
+		t.Errorf("should not show 'ludus deploy session' when withSession is true, got: %q", output)
+	}
+}
+
+func TestPrintNextStepWithDeploySkipped(t *testing.T) {
+	// Verify printNextStep when deploy is skipped shows simpler guidance
+	origSkip := skipDeploy
+	origSession := withSession
+	t.Cleanup(func() {
+		skipDeploy = origSkip
+		withSession = origSession
+	})
+
+	skipDeploy = true
+	withSession = false
+
+	output := captureStdout(printNextStep)
+	if !strings.Contains(output, "Pipeline complete") {
+		t.Errorf("expected 'Pipeline complete', got: %q", output)
+	}
+	// When deploy is skipped, should not show deploy guidance
+	if strings.Contains(output, "ludus deploy session") || strings.Contains(output, "ludus connect") {
+		t.Errorf("should not show deploy/connect guidance when skipDeploy is true, got: %q", output)
 	}
 }
